+++ lib/functions/display.php
@@ -82,17 +82,27 @@ function mai_content_filter_shortcodes( $content ) {
  * @return  void
  */
 function mai_do_featured_image( $size = 'featured' ) {
-	echo '<div class="featured-image">';
-		echo genesis_get_image( array(
-			'format' => 'html',
-			'size'   => $size,
-			'attr'   => array( 'class' => 'wp-post-image' )
-			));
-	echo '</div>';
 
-	$caption = get_post( get_post_thumbnail_id() )->post_excerpt;
+	$image = genesis_get_image( array(
+		'format' => 'html',
+		'size'   => $size,
+		'attr'   => array( 'class' => 'wp-post-image' )
+	) );
+
+	if ( ! $image ) {
+		return;
+	}
+
+	printf( '<div class="featured-image">%s</div>', $image );
+
+	$attachment = get_post( get_post_thumbnail_id() );
+	if ( ! $attachment ) {
+		return;
+	}
+
+	$caption = $attachment->post_excerpt;
 	if ( $caption ) {
-		echo '<span class="image-caption">' . $caption . '</span>';
+		printf( '<span class="image-caption">%s</span>', $caption );
 	}
 }