+++ lib/shortcodes.php
@@ -126,51 +126,41 @@ function mai_get_col_one_whole_shortcode( $atts, $content = null ) {
 }
 
 /**
- * Add utility classes to the default gallery shortcode wrap.
- *
- * @since   1.4.2
- *
- * @param   string   $output  Default CSS styles and opening HTML div container
- *                            for the gallery shortcode output.
- *
- * @return  string  The gallery HTML.
- */
-add_filter( 'gallery_style', 'mai_gallery_style' );
-function mai_gallery_style( $output ) {
-	return str_replace( "class='", "class='row gutter-md ", $output );
-}
-
-/**
  * Add utility classes to the default gallery shortcode.
  *
- * @since   1.3.8
+ * @since   1.4.3
  *
- * @param string $output   The gallery output. Default empty.
- * @param array  $atts     Attributes of the gallery shortcode.
- * @param int    $instance Unique numeric ID of this gallery shortcode instance.
+ * @param   string        $output  Shortcode output.
+ * @param   string        $tag     Shortcode name.
+ * @param   array|string  $attr    Shortcode attributes array or empty string.
+ * @param   array         $m       Regular expression match array.
  *
  * @return  string  The gallery HTML.
  */
-add_filter( 'post_gallery', 'mai_post_gallery', 10, 3 );
-function mai_post_gallery( $output, $atts, $instance ) {
+add_filter( 'do_shortcode_tag', 'mai_gallery_shortcode_tag', 10, 4 );
+function mai_gallery_shortcode_tag( $output, $tag, $atts, $m ) {
 
-	// Remove filter to avoid infinite loop.
-	remove_filter( 'post_gallery', 'mai_post_gallery', 10, 3 );
+	// Bail if not a gallery.
+	if ( 'gallery' !== $tag ) {
+		return $output;
+	}
 
-	// Make sure we have a columns value.
+	// Bail if not a default gallery. This fixes compatibility with Jetpack galleries.
+	if ( isset( $atts['type'] ) && ! in_array( $atts['type'], array( 'default', 'thumbnails' ) ) ) {
+		return $output;
+	}
+
+	// Row classes.
+	$output = str_replace( "class='gallery galleryid", "class='row gutter-md gallery galleryid", $output );
+
+	// Make sure we have a columns value. Would not be set if 3 as per WP core.
 	$atts = wp_parse_args( $atts, array( 'columns' => 3 ) );
 
-	// Get the full gallery markup.
-	$output = gallery_shortcode( $atts );
-
-	// Build our new entry/item classes.
+	// Build column classes.
 	$classes = 'gallery-item col';
 	$classes = mai_add_classes( mai_get_col_classes_by_columns( $atts['columns'] ), $classes );
 	$classes = mai_add_classes( 'bottom-xs-md', $classes );
 	$output  = str_replace( 'gallery-item', $classes, $output );
-
-	// Add filter back incase there is another gallery.
-	add_filter( 'post_gallery', 'mai_post_gallery', 10, 3 );
 
 	return $output;
 }